requirements.txt
requests
pillow
pandas
tqdm
python-dotenv
openpyxl